Fixes to mouse/keyboard/touchscreen input.
A date variable is updated every time there is aforementioned interaction.

Every minute, we see how long ago the last interaction was. If greater than predefined timeout, we send a logout call to the API then remove the cookie.

Otherwise, every minute a PING is sent to let the server know we're still doing something.

So now there are 4 safeguards:

1.) Server check every time a request comes in
2.) Every navigation sends a PING to check if the server has timed out the session.
3.) The cookie is set to expire in X minutes; reset by aforementioned interaction
4.) Ping check every minute.